Moved ViewModels back to UI SEtup UI project to support testing Add UI project to testing Cleanup mac/win compiling differently.
