--- src/electionguard-ui/ElectionGuard.UI.Lib/ViewModels/SettingsViewModel.cs
@@ -1,9 +1,0 @@
-﻿namespace ElectionGuard.UI.Lib.ViewModels
-{
-    public class SettingsViewModel : BaseViewModel
-    {
-        public SettingsViewModel(IServiceProvider serviceProvider) : base(null, serviceProvider)
-        {
-        }
-    }
-}
