+++ src/electionguard-ui/ElectionGuard.UI.Test/ViewModels/ViewKeyCeremonyViewModelTest.cs
@@ -1,4 +1,4 @@
-﻿using ElectionGuard.UI.Lib.ViewModels;
+﻿using ElectionGuard.UI.ViewModels;
 using NSubstitute;
 using Shouldly;
