+++ src/electionguard-ui/ElectionGuard.UI/AppShell.xaml.cs
@@ -1,6 +1,7 @@
 using ElectionGuard.UI.Lib.Services;
 using ElectionGuard.UI.Services;
 using ElectionGuard.UI.Lib.Extensions;
+using CommunityToolkit.Mvvm.DependencyInjection;
 
 namespace ElectionGuard.UI;
 
@@ -10,7 +11,7 @@ public partial class AppShell
     {
         InitializeComponent();
         
-        var navigationService = EgServiceProvider.Current.GetInstance<INavigationService>();
+        var navigationService = Ioc.Default.GetInstance<INavigationService> ();
         navigationService.RegisterRoutes();
     }
 }
