+++ src/electionguard-ui/ElectionGuard.UI/GlobalUsings.cs
@@ -1,7 +1,9 @@
 global using CommunityToolkit.Mvvm.ComponentModel;
+global using System.Collections.ObjectModel;
 
 global using ElectionGuard.UI.Helpers;
 global using ElectionGuard.UI.Lib.Models;
 global using ElectionGuard.UI.Resx;
 global using ElectionGuard.UI.Views;
-global using ElectionGuard.UI.Lib.ViewModels;
+global using ElectionGuard.UI.ViewModels;
+global using ElectionGuard.UI.Lib.Services;
