+++ src/electionguard-ui/ElectionGuard.UI/ViewModels/SettingsViewModel.cs
@@ -1,0 +1,8 @@
+﻿namespace ElectionGuard.UI.ViewModels;
+
+public class SettingsViewModel : BaseViewModel
+{
+    public SettingsViewModel(IServiceProvider serviceProvider) : base(null, serviceProvider)
+    {
+    }
+}
