+++ src/electionguard-ui/ElectionGuard.UI/Views/AdminHomePage.xaml
@@ -5,7 +5,7 @@
                        xmlns:helper="clr-namespace:ElectionGuard.UI.Helpers"
                        xmlns:model="clr-namespace:ElectionGuard.UI.Lib.Models;assembly=ElectionGuard.UI.Lib"
                        xmlns:controls="clr-namespace:ElectionGuard.UI.Controls"
-                       xmlns:viewModels="clr-namespace:ElectionGuard.UI.Lib.ViewModels;assembly=ElectionGuard.UI.Lib"
+                       xmlns:viewModels="clr-namespace:ElectionGuard.UI.ViewModels"
                        xmlns:views="using:ElectionGuard.UI.Views"
                        x:DataType="viewModels:AdminHomeViewModel"
                        ControlTemplate="{StaticResource PageTemplate}">
