+++ src/electionguard-ui/ElectionGuard.UI/Views/GuardianHomePage.xaml
@@ -1,15 +1,15 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <views:ContentPageBase x:TypeArguments="viewModels:GuardianHomeViewModel"
-    xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-                                                        xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-                                                        x:Class="ElectionGuard.UI.Views.GuardianHomePage"
-                                                        xmlns:helper="clr-namespace:ElectionGuard.UI.Helpers"
-                                                        xmlns:model="clr-namespace:ElectionGuard.UI.Lib.Models;assembly=ElectionGuard.UI.Lib"
-                                                        xmlns:viewModels="clr-namespace:ElectionGuard.UI.Lib.ViewModels;assembly=ElectionGuard.UI.Lib"
-                                                        xmlns:controls="clr-namespace:ElectionGuard.UI.Controls"
-                                                        xmlns:views="using:ElectionGuard.UI.Views"
-                                                        x:DataType="viewModels:GuardianHomeViewModel"
-                                                        ControlTemplate="{StaticResource PageTemplate}">
+                       xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+                        xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+                        x:Class="ElectionGuard.UI.Views.GuardianHomePage"
+                        xmlns:helper="clr-namespace:ElectionGuard.UI.Helpers"
+                        xmlns:model="clr-namespace:ElectionGuard.UI.Lib.Models;assembly=ElectionGuard.UI.Lib"
+                        xmlns:viewModels="clr-namespace:ElectionGuard.UI.ViewModels"
+                        xmlns:controls="clr-namespace:ElectionGuard.UI.Controls"
+                        xmlns:views="using:ElectionGuard.UI.Views"
+                        x:DataType="viewModels:GuardianHomeViewModel"
+                        ControlTemplate="{StaticResource PageTemplate}">
     <VerticalStackLayout Style="{StaticResource FullWidthPage}">
         <Grid ColumnDefinitions="*,*">
             <VerticalStackLayout Grid.Column="0" >
