+++ src/electionguard-ui/ElectionGuard.UI/Views/LoginPage.xaml
@@ -3,7 +3,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:helper="clr-namespace:ElectionGuard.UI.Helpers"
              x:Class="ElectionGuard.UI.Views.LoginPage"
-             xmlns:viewModels="clr-namespace:ElectionGuard.UI.Lib.ViewModels;assembly=ElectionGuard.UI.Lib"
+             xmlns:viewModels="clr-namespace:ElectionGuard.UI.ViewModels"
              x:DataType="viewModels:LoginViewModel"
              ControlTemplate="{StaticResource PageTemplate}">
     <VerticalStackLayout Style="{StaticResource NarrowCenteredPage}" Margin="0,100,0,0">
